@@ -115,14 +115,19 @@ subprojects {
 
     repositories {
         mavenCentral()
+        maven {
+            url "http://maven.java.net/content/repositories/snapshots"
+        }
     }
 
-    sourceCompatibility = 1.8
+    sourceCompatibility = JavaVersion.VERSION_1_9
     compileJava.options.encoding = 'UTF-8'
 
     dependencies {
         compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.21'
-        compile group: 'com.google.guava', name: 'guava', version: '23.0'
+        compile (group: 'com.google.guava', name: 'guava', version: '25.0-jre'){
+            exclude module: "jsr305" // module conflict: https://blog.codefx.org/java/jsr-305-java-9/, https://github.com/google/guava/issues/2960
+        }
         testCompile group: 'junit', name: 'junit', version: '4.12'
         testCompile group: 'org.mockito', name: 'mockito-all', version: '1.9.5'
     }
