@@ -1,40 +1,28 @@
 package de.factoryfx.data.attribute;
 
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.List;
+import java.util.*;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import javafx.collections.FXCollections;
-import javafx.collections.ListChangeListener;
-import javafx.collections.ObservableList;
+import de.factoryfx.data.Data;
 
-public class ValueListAttribute<T, A extends Attribute<List<T>,A>> extends ImmutableValueAttribute<List<T>,A> implements Collection<T> {
+public class ValueListAttribute<T, A extends Attribute<List<T>,A>> extends ImmutableValueAttribute<List<T>,A> implements List<T> {
     private final Class<T> itemType;
-    private final ObservableList<T> observableValue;
 
     public ValueListAttribute(Class<T> itemType) {
         super(null);
-        this.itemType=itemType;
-        observableValue = FXCollections.observableArrayList();
-        value=observableValue;
-
-        observableValue.addListener((ListChangeListener<T>) c -> {
-            updateListeners(get());
-        });
+        this.itemType = itemType;
+        this.value = new ArrayList<>();
     }
 
-//    /**
-//     *
-//     * @param attributeMetadata AttributeMetadata
-//     * @param itemType generics workaround, if itemType is generic the correct constructor don't work
-//     */
-//    @SuppressWarnings("unchecked")
-//    public ValueListAttribute(AttributeMetadata attributeMetadata,Class itemType) {
-//        this((Class<T>)itemType,attributeMetadata);
-//    }
+    private void afterModify(){
+        if (listeners!=null) {
+            for (AttributeChangeListener<List<T>, A> listener : listeners) {
+                listener.changed(ValueListAttribute.this, this.value);
+            }
+        }
+    }
 
     @JsonCreator
     protected ValueListAttribute() {
@@ -43,91 +31,156 @@ public class ValueListAttribute<T, A extends Attribute<List<T>,A>> extends Immut
 
     @Override
     public String getDisplayText() {
-        return new CollectionAttributeUtil<>(get(), Object::toString).getDisplayText();
+        return new CollectionAttributeUtil<>(this.value, Object::toString).getDisplayText();
     }
 
     @Override
     public AttributeTypeInfo internal_getAttributeType() {
-        return new AttributeTypeInfo(ObservableList.class,null,null,itemType, AttributeTypeInfo.AttributeTypeCategory.COLLECTION);
+        return new AttributeTypeInfo(Set.class,null,null,itemType, AttributeTypeInfo.AttributeTypeCategory.COLLECTION);
     }
 
     //** set list only take the list items not the list itself, (to simplify ChangeListeners)*/
     @Override
     public void set(List<T> value) {
         if (value==null){//workaround for jackson
-            observableValue.clear();
+            this.value.clear();
         } else {
-            observableValue.setAll(value);
+            this.value.clear();
+            this.value.addAll(value);
         }
     }
 
     public List<T> filtered(Predicate<T> predicate) {
-        return get().stream().filter(predicate).collect(Collectors.toList());
+        return this.value.stream().filter(predicate).collect(Collectors.toList());
     }
 
     @Override
     public int size() {
-        return get().size();
+        return this.value.size();
     }
 
     @Override
     public boolean isEmpty() {
-        return get().isEmpty();
+        return this.value.isEmpty();
     }
 
     @Override
     public boolean contains(Object o) {
-        return get().contains(o);
+        return this.value.contains(o);
     }
 
     @Override
     public Iterator<T> iterator() {
-        return get().iterator();
+        return this.value.iterator();
     }
 
     @Override
     public Object[] toArray() {
-        return get().toArray();
+        return this.value.toArray();
     }
 
     @Override
     public <T1> T1[] toArray(T1[] a) {
-        return get().toArray(a);
+        return this.value.toArray(a);
     }
 
     @Override
     public boolean add(T t) {
-        return get().add(t);
+        boolean add = this.value.add(t);
+        afterModify();
+        return add;
     }
 
     @Override
     public boolean remove(Object o) {
-        return get().remove(o);
+        boolean remove = this.value.remove(o);
+        afterModify();
+        return remove;
     }
 
     @Override
     public boolean containsAll(Collection<?> c) {
-        return get().containsAll(c);
+        return this.value.containsAll(c);
     }
 
     @Override
     public boolean addAll(Collection<? extends T> c) {
-        return get().addAll(c);
+        boolean b = this.value.addAll(c);
+        afterModify();
+        return b;
+    }
+
+    @Override
+    public boolean addAll(int index, Collection<? extends T> c) {
+        boolean b = this.value.addAll(index, c);
+        afterModify();
+        return b;
     }
 
     @Override
     public boolean removeAll(Collection<?> c) {
-        return get().removeAll(c);
+        boolean b = this.value.removeAll(c);
+        afterModify();
+        return b;
     }
 
     @Override
     public boolean retainAll(Collection<?> c) {
-        return get().retainAll(c);
+        return this.value.retainAll(c);
     }
 
     @Override
     public void clear() {
-        get().clear();
+        this.value.clear();
+        afterModify();
+    }
+
+    @Override
+    public T get(int index) {
+        return this.value.get(index);
+    }
+
+    @Override
+    public T set(int index, T element) {
+        return this.value.set(index,element);
+    }
+
+    @Override
+    public void add(int index, T element) {
+        this.value.add(index,element);
+        afterModify();
+    }
+
+    @Override
+    public T remove(int index) {
+        T remove = this.value.remove(index);
+        afterModify();
+        return remove;
+    }
+
+    @Override
+    public int indexOf(Object o) {
+        return this.value.indexOf(o);
+    }
+
+    @Override
+    public int lastIndexOf(Object o) {
+        return this.value.lastIndexOf(o);
+    }
+
+    @Override
+    public ListIterator<T> listIterator() {
+        return this.value.listIterator();
+    }
+
+    @Override
+    public ListIterator<T> listIterator(int index) {
+        return this.value.listIterator(index);
+    }
+
+    @Override
+    public List<T> subList(int fromIndex, int toIndex) {
+        return this.value.subList(fromIndex,toIndex);
     }
 
 }
