@@ -3,9 +3,8 @@ package de.factoryfx.data.attribute;
 import java.util.*;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableSet;
-import javafx.collections.SetChangeListener;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 
 public abstract class ValueSetAttribute<T,A extends Attribute<Set<T>,A>> extends ImmutableValueAttribute<Set<T>,A> implements Set<T> {
     private final Class<T> itemType;
@@ -13,12 +12,7 @@ public abstract class ValueSetAttribute<T,A extends Attribute<Set<T>,A>> extends
     public ValueSetAttribute(Class<T> itemType) {
         super(null);
         this.itemType = itemType;
-        final ObservableSet<T> observableSet = FXCollections.observableSet(new HashSet<>());
-        value= observableSet;
-
-        observableSet.addListener((SetChangeListener<T>) change -> {
-            updateListeners(get());
-        });
+        this.value= new HashSet<>();
     }
 
     @JsonCreator
@@ -28,83 +22,107 @@ public abstract class ValueSetAttribute<T,A extends Attribute<Set<T>,A>> extends
 
     @Override
     public String getDisplayText() {
-        return new CollectionAttributeUtil<>(get(), Object::toString).getDisplayText();
+        return new CollectionAttributeUtil<>(this.value, Object::toString).getDisplayText();
     }
 
     @Override
     public AttributeTypeInfo internal_getAttributeType() {
-        return new AttributeTypeInfo(ObservableSet.class,null,null,itemType, AttributeTypeInfo.AttributeTypeCategory.COLLECTION);
+        return new AttributeTypeInfo(Set.class,null,null,itemType, AttributeTypeInfo.AttributeTypeCategory.COLLECTION);
     }
 
     //** set list only take the list items not the list itself, (to simplify ChangeListeners)*/
     @Override
     public void set(Set<T> value) {
-        this.get().clear();
-        this.get().addAll(value);
+        this.value.clear();
+        this.value.addAll(value);
+    }
+
+    //FIXME useful but breaks jackson, try to find workaround
+//    @JsonIgnore
+//    @Override
+//    public Set<T> get() {
+//        return this;
+//    }
+
+    private void afterModify(){
+        if (listeners!=null) {
+            for (AttributeChangeListener<Set<T>, A> listener : listeners) {
+                listener.changed(ValueSetAttribute.this, this.value);
+            }
+        }
     }
 
     @Override
     public int size() {
-        return get().size();
+        return this.value.size();
     }
 
     @Override
     public boolean isEmpty() {
-        return get().isEmpty();
+        return this.value.isEmpty();
     }
 
     @Override
     public boolean contains(Object o) {
-        return get().contains(o);
+        return this.value.contains(o);
     }
 
     @Override
     public Iterator<T> iterator() {
-        return get().iterator();
+        return this.value.iterator();
     }
 
     @Override
     public Object[] toArray() {
-        return get().toArray();
+        return this.value.toArray();
     }
 
     @Override
     public <T1> T1[] toArray(T1[] a) {
-        return get().toArray(a);
+        return this.value.toArray(a);
     }
 
     @Override
     public boolean add(T t) {
-        return get().add(t);
+        boolean add = this.value.add(t);
+        afterModify();
+        return add;
     }
 
     @Override
     public boolean remove(Object o) {
-        return get().remove(o);
+        boolean remove = this.value.remove(o);
+        afterModify();
+        return remove;
     }
 
     @Override
     public boolean containsAll(Collection<?> c) {
-        return get().containsAll(c);
+        return this.value.containsAll(c);
     }
 
     @Override
     public boolean addAll(Collection<? extends T> c) {
-        return get().addAll(c);
+        boolean b = value.addAll(c);
+        afterModify();
+        return b;
     }
 
     @Override
     public boolean retainAll(Collection<?> c) {
-        return get().retainAll(c);
+        return this.value.retainAll(c);
     }
 
     @Override
     public boolean removeAll(Collection<?> c) {
-        return get().removeAll(c);
+        boolean b = this.value.removeAll(c);
+        afterModify();
+        return b;
     }
 
     @Override
     public void clear() {
-        get().clear();
+        this.value.clear();
+        afterModify();
     }
 }
