@@ -146,7 +146,7 @@ public class ViewListReferenceAttributeTest {
         }
 
         root = root.internal().prepareUsableCopy();
-        root.ref.get().view.setRunlaterExecutorForTest(runnable -> runnable.run());
+        root.ref.get().view.setRunlaterExecutor(runnable -> runnable.run());
 
         ArrayList<String> calls=new ArrayList<>();
         root.ref.get().view.internal_addListener((attribute, value1) -> {
