@@ -2,7 +2,7 @@ package de.factoryfx.docu.dynamicwebserver;
 
 import de.factoryfx.factory.SimpleFactoryBase;
 import de.factoryfx.factory.atrribute.FactoryReferenceAttribute;
-import de.factoryfx.server.rest.server.JettyServer;
+import de.factoryfx.jetty.JettyServer;
 
 public class RootFactory extends SimpleFactoryBase<JettyServer,Void,RootFactory> {
     public final FactoryReferenceAttribute<JettyServer,Main.DynamicWebserver> server=new FactoryReferenceAttribute<>(Main.DynamicWebserver.class).labelText("server");
