@@ -8,22 +8,22 @@ import de.factoryfx.example.factory.OrderCollector;
 import de.factoryfx.example.factory.ShopFactory;
 import de.factoryfx.factory.builder.FactoryTreeBuilder;
 import de.factoryfx.factory.builder.Scope;
-import de.factoryfx.javafx.editor.DataEditorFactory;
-import de.factoryfx.javafx.editor.attribute.AttributeEditorBuilderFactory;
-import de.factoryfx.javafx.editor.attribute.EditorBuildersListFactory;
-import de.factoryfx.javafx.stage.DefaultStageFactory;
-import de.factoryfx.javafx.util.LongRunningActionExecutorFactory;
-import de.factoryfx.javafx.util.UniformDesignFactory;
-import de.factoryfx.javafx.view.ViewDescriptionFactory;
-import de.factoryfx.javafx.view.ViewFactory;
-import de.factoryfx.javafx.view.container.ViewsDisplayWidgetFactory;
-import de.factoryfx.javafx.view.factoryviewmanager.FactoryEditManagerFactory;
-import de.factoryfx.javafx.view.factoryviewmanager.FactoryEditViewFactory;
-import de.factoryfx.javafx.view.menu.ViewMenuFactory;
-import de.factoryfx.javafx.view.menu.ViewMenuItemFactory;
-import de.factoryfx.javafx.widget.diffdialog.DiffDialogBuilderFactory;
-import de.factoryfx.server.rest.client.MicroserviceRestClientFactory;
-import de.factoryfx.server.rest.client.RestClientFactory;
+import de.factoryfx.javafx.factory.editor.DataEditorFactory;
+import de.factoryfx.javafx.factory.editor.attribute.AttributeEditorBuilderFactory;
+import de.factoryfx.javafx.factory.editor.attribute.EditorBuildersListFactory;
+import de.factoryfx.javafx.factory.stage.DefaultStageFactory;
+import de.factoryfx.javafx.factory.util.LongRunningActionExecutorFactory;
+import de.factoryfx.javafx.factory.util.UniformDesignFactory;
+import de.factoryfx.javafx.factory.view.ViewDescriptionFactory;
+import de.factoryfx.javafx.factory.view.ViewFactory;
+import de.factoryfx.javafx.factory.view.container.ViewsDisplayWidgetFactory;
+import de.factoryfx.javafx.factory.view.factoryviewmanager.FactoryEditManagerFactory;
+import de.factoryfx.javafx.factory.view.factoryviewmanager.FactoryEditViewFactory;
+import de.factoryfx.javafx.factory.view.menu.ViewMenuFactory;
+import de.factoryfx.javafx.factory.view.menu.ViewMenuItemFactory;
+import de.factoryfx.javafx.factory.widget.factory.diffdialog.DiffDialogBuilderFactory;
+import de.factoryfx.microservice.rest.client.MicroserviceRestClientFactory;
+import de.factoryfx.util.rest.client.RestClientFactory;
 import javafx.geometry.Rectangle2D;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
