@@ -11,9 +11,9 @@ import de.factoryfx.factory.FactoryManager;
 import de.factoryfx.factory.exception.AllOrNothingFactoryExceptionHandler;
 import de.factoryfx.factory.exception.LoggingFactoryExceptionHandler;
 import de.factoryfx.factory.exception.RethrowingFactoryExceptionHandler;
+import de.factoryfx.jetty.HttpServerConnectorFactory;
+import de.factoryfx.microservice.rest.MicroserviceResourceFactory;
 import de.factoryfx.server.Microservice;
-import de.factoryfx.server.rest.MicroserviceResourceFactory;
-import de.factoryfx.server.rest.server.HttpServerConnectorFactory;
 import javafx.application.Application;
 import javafx.stage.Stage;
 import org.slf4j.LoggerFactory;
