@@ -7,11 +7,12 @@ import de.factoryfx.factory.PolymorphicFactoryBase;
 import de.factoryfx.factory.atrribute.FactoryReferenceListAttribute;
 import de.factoryfx.server.user.User;
 import de.factoryfx.server.user.UserManagement;
+import de.factoryfx.server.user.persistent.UserFactory;
 
 import java.util.HashSet;
 
 public class PersistentUserManagementFactory<V,R extends FactoryBase<?,V,R>> extends PolymorphicFactoryBase<UserManagement,V,R>  {
-    public final FactoryReferenceListAttribute<User,UserFactory<V,R>> users = new FactoryReferenceListAttribute<User,UserFactory<V,R>>().setupUnsafe(UserFactory.class).en("users").de("Benutzer").userNotSelectable();
+    public final FactoryReferenceListAttribute<User, UserFactory<V,R>> users = new FactoryReferenceListAttribute<User, UserFactory<V,R>>().setupUnsafe(UserFactory.class).en("users").de("Benutzer").userNotSelectable();
 
     @Override
     public PersistentUserManagement createImpl() {
