@@ -1,6 +1,7 @@
 package de.factoryfx.javafx.javascript.editor.attribute.visualisation;
 
-import de.factoryfx.javafx.editor.attribute.ValueAttributeEditorVisualisation;
+
+import de.factoryfx.javafx.data.editor.attribute.ValueAttributeEditorVisualisation;
 import de.factoryfx.javascript.data.attributes.types.JavascriptAttribute;
 import de.factoryfx.javascript.data.attributes.types.Javascript;
 import javafx.beans.property.SimpleObjectProperty;
