@@ -7,12 +7,12 @@ import java.util.function.Consumer;
 import de.factoryfx.data.Data;
 import de.factoryfx.data.attribute.Attribute;
 import de.factoryfx.data.jackson.ObjectMapperBuilder;
-import de.factoryfx.javafx.editor.attribute.AttributeEditor;
-import de.factoryfx.javafx.editor.attribute.AttributeEditorBuilder;
-import de.factoryfx.javafx.editor.attribute.builder.SingleAttributeEditorBuilder;
-import de.factoryfx.javafx.editor.data.DataEditor;
+import de.factoryfx.javafx.data.editor.attribute.AttributeEditor;
+import de.factoryfx.javafx.data.editor.attribute.AttributeEditorBuilder;
+import de.factoryfx.javafx.data.editor.attribute.builder.SingleAttributeEditorBuilder;
+import de.factoryfx.javafx.data.editor.data.DataEditor;
+import de.factoryfx.javafx.data.util.UniformDesign;
 import de.factoryfx.javafx.javascript.editor.attribute.visualisation.JavascriptAttributeVisualisation;
-import de.factoryfx.javafx.util.UniformDesign;
 import de.factoryfx.javascript.data.attributes.types.Javascript;
 import de.factoryfx.javascript.data.attributes.types.JavascriptAttribute;
 import javafx.application.Application;
